Refactored Hardcoded Password Equals Detectors
